Create DataBaseQuery
This is a PHP script which will take input from a form and on the basis of 
that query a database and then display the data in a HTML table. The form
will offer the user a choice of fields to sort the data by as well as whether
they want the data to be sorted in ascending or descending order.